azalea_chat::text_component
Function legacy_color_code_to_text_component

pub fn legacy_color_code_to_text_component(
    legacy_color_code: &str,
) -> TextComponent

Convert a legacy color code string into a FormattedText Technically in Minecraft this is done when displaying the text, but AFAIK it’s the same as just doing it in TextComponent